Removed redundant component
